requirements.txt
pandas
sqlalchemy
psycopg2-binary
matplotlib
seaborn
python-dotenv
jupyter
